requirements.txt
openai
flask
PyMuPDF
python-docx
PyPDF2
python-dotenv
langchain
word2number
tiktoken
faiss-cpu